requirements.txt
tweety-ns>=2.4.0
python-telegram-bot>=21.0
python-dotenv>=1.0.0
yt-dlp>=2024.0
httpx>=0.27.0
